+++ src/Covenant/Internal/Unification.hs
@@ -1,5 +1,4 @@
 {-# LANGUAGE CPP #-}
-{-# LANGUAGE MultiWayIf #-}
 
 module Covenant.Internal.Unification
   ( TypeAppError (..),
@@ -21,7 +20,7 @@ import Data.Foldable (foldl')
 #endif
 import Control.Monad.Except (MonadError, catchError, throwError)
 import Control.Monad.Reader (MonadReader, ReaderT (runReaderT), ask)
-import Covenant.Data (DatatypeInfo, mkDatatypeInfo)
+import Covenant.Data (DatatypeInfo)
 import Covenant.Index (Index, intCount, intIndex)
 import Covenant.Internal.Rename (RenameError, renameDatatypeInfo)
 import Covenant.Internal.Type
@@ -30,11 +29,8 @@ import Covenant.Internal.Type
     CompT (CompT),
     CompTBody (CompTBody),
     Renamed (Rigid, Unifiable, Wildcard),
-    TyName (TyName),
+    TyName,
     ValT (Abstraction, BuiltinFlat, Datatype, ThunkT),
-    byteStringBaseFunctor,
-    naturalBaseFunctor,
-    negativeBaseFunctor,
   )
 import Data.Kind (Type)
 import Data.Map (Map)
@@ -44,7 +40,6 @@ import Data.Maybe (fromJust, mapMaybe)
 import Data.Set (Set)
 import Data.Set qualified as Set
 import Data.Text (Text)
-import Data.Text qualified as Text
 import Data.Vector (Vector)
 import Data.Vector qualified as Vector
 import Data.Vector.NonEmpty (NonEmptyVector)
@@ -114,42 +109,13 @@ runUnifyM tyDict (UnifyM act) = runReaderT act tyDict
 lookupDatatypeInfo ::
   TyName ->
   UnifyM (DatatypeInfo Renamed)
-lookupDatatypeInfo tn@(TyName rawTyName) =
+lookupDatatypeInfo tn =
   ask >>= \tyDict -> case preview (ix tn) tyDict of
-    Nothing -> checkForBaseFunctor tyDict
+    Nothing -> throwError . NoDatatypeInfo $ tn
     Just dti -> renamedToUnify . renameDatatypeInfo $ dti
   where
-    checkForBaseFunctor :: Map TyName (DatatypeInfo AbstractTy) -> UnifyM (DatatypeInfo Renamed)
-    checkForBaseFunctor tyDict = case Text.stripPrefix "#" rawTyName of
-      Nothing -> throwError . NoDatatypeInfo $ tn
-      Just rawTyNameStub ->
-        if
-          -- Note (Koz, 12/08/2025): None of these specific cases should _ever_
-          -- fail. Thus, `fromRight` is safe here.
-          | rawTyNameStub == "Natural" ->
-              renamedToUnify . renameDatatypeInfo . fromRight . mkDatatypeInfo $ naturalBaseFunctor
-          | rawTyNameStub == "Negative" ->
-              renamedToUnify . renameDatatypeInfo . fromRight . mkDatatypeInfo $ negativeBaseFunctor
-          | rawTyNameStub == "ByteString" ->
-              renamedToUnify . renameDatatypeInfo . fromRight . mkDatatypeInfo $ byteStringBaseFunctor
-          -- We have something that _looks_ like a base functor, but not a
-          -- special builtin case. We thus need to ask the environment for the
-          -- recursive type it stands for, if it exists.
-          | otherwise -> do
-              let standinTyName = TyName rawTyNameStub
-              case preview (ix standinTyName) tyDict of
-                -- Now we have _truly_ missed.
-                Nothing -> throwError . NoDatatypeInfo $ tn
-                Just dti -> case view #baseFunctor dti of
-                  Nothing -> throwError . NoDatatypeInfo $ tn
-                  -- Since this is generated, it can't fail to rename
-                  Just (bfDd, _) -> renamedToUnify . renameDatatypeInfo . fromRight . mkDatatypeInfo $ bfDd
     renamedToUnify :: Either RenameError (DatatypeInfo Renamed) -> UnifyM (DatatypeInfo Renamed)
     renamedToUnify = either (throwError . DatatypeInfoRenameFailed tn) pure
-    fromRight :: forall a b. (Show a) => Either a b -> b
-    fromRight = \case
-      Left err -> error . show $ err
-      Right x -> x
 
 lookupBBForm :: TyName -> UnifyM (ValT Renamed)
 lookupBBForm tn =