+++ src/Covenant/Test.hs
@@ -87,7 +87,7 @@ import Covenant.ASG (ASGEnv (ASGEnv), ASGNode, CovenantError (TypeError), Covena
 import Covenant.Data
   ( DatatypeInfo,
     mkDatatypeInfo,
-    noPhantomTyVars,
+    noPhantomTyVars, primBaseFunctorInfos,
   )
 import Covenant.DeBruijn (DeBruijn (Z), asInt)
 import Covenant.Index
@@ -371,7 +371,7 @@ failLeft = either (assertFailure . show) pure
 -- @since 1.1.0
 tyAppTestDatatypes :: M.Map TyName (DatatypeInfo AbstractTy)
 tyAppTestDatatypes =
-  foldl' (\acc decl -> M.insert (view #datatypeName decl) (unsafeMkDatatypeInfo decl) acc) M.empty testDatatypes
+  foldl' (\acc decl -> unsafeMkDatatypeInfo decl <> acc) primBaseFunctorInfos testDatatypes
   where
     unsafeMkDatatypeInfo d = case mkDatatypeInfo d of
       Left err -> error (show err)